+++ apps/govern/components/Contract.tsx
@@ -5,7 +5,8 @@ import styled from 'styled-components';
 import { mainnet } from 'viem/chains';
 import { useEnsName } from 'wagmi';
 
-import { UNICODE_SYMBOLS } from 'common-util/constants/unicode';
+import { UNICODE_SYMBOLS } from 'libs/util-constants/src';
+
 import { EXPLORE_URLS } from 'common-util/constants/urls';
 import { CHAIN_NAMES, getAddressFromBytes32, truncateAddress } from 'common-util/functions';
 import { useContractParams } from 'hooks/index';
