--- apps/govern/pages/index.tsx
@@ -1,3 +1,0 @@
-import { ContractsPage } from '../components/Contracts';
-
-export default ContractsPage;
