+++ libs/util-constants/README.md
@@ -1,0 +1,3 @@
+# util-constants
+
+This library was generated with [Nx](https://nx.dev).
