+++ libs/util-constants/src/lib/ipfs.ts
@@ -1,0 +1,2 @@
+export const GATEWAY_URL = 'https://gateway.autonolas.tech/ipfs/';
+export const HASH_PREFIX = 'f01701220';
